tests/requirements.txt
ruamel.yaml
requests
urllib3
git+http://github.com/Datera/python-sdk
